[Bronze V] Title: 사파리월드, Time: 0 ms, Memory: 2020 KB -BaekjoonHub

+++ 백준/Bronze/2420. 사파리월드/사파리월드.cc
@@ -2,7 +2,7 @@
 #include <iostream>
 using namespace std;
 int main(void) {
-    long long n, m;
+    long n, m;
     cin >> n >> m;
     cout << abs(n - m);
     return 0;
